requirements.txt
gradio
ultralytics
opencv-python
tensorflow
numpy
Pillow
scikit-learn
gtts